+++ src/main/java/es/in2/issuer/domain/service/JWTService.java
@@ -1,5 +1,6 @@
 package es.in2.issuer.domain.service;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.nimbusds.jose.JWSObject;
 import com.nimbusds.jose.Payload;
 import com.nimbusds.jwt.SignedJWT;
@@ -7,9 +8,9 @@ import reactor.core.publisher.Mono;
 
 public interface JWTService {
 
-    String generateJWT(String payload);
 
     Mono<Boolean> validateJwtSignatureReactive(JWSObject jwsObject);
+    Mono<JsonNode> parseJwtVCAsJsonNode(String jwt);
 
     SignedJWT parseJWT(String jwt);