+++ src/main/java/es/in2/issuer/domain/service/VerifierValidationService.java
@@ -1,0 +1,8 @@
+package es.in2.issuer.domain.service;
+
+import reactor.core.publisher.Mono;
+
+public interface VerifierValidationService {
+
+    Mono<Void> verifyToken(String accessToken);
+}