+++ src/main/java/es/in2/issuer/infrastructure/config/AppConfig.java
@@ -4,7 +4,6 @@ import es.in2.issuer.infrastructure.config.adapter.ConfigAdapter;
 import es.in2.issuer.infrastructure.config.adapter.factory.ConfigAdapterFactory;
 import es.in2.issuer.infrastructure.config.properties.ApiProperties;
 import es.in2.issuer.infrastructure.config.properties.IssuerUiProperties;
-import es.in2.issuer.infrastructure.config.properties.TrustServiceProviderForCertificationsProperties;
 import es.in2.issuer.infrastructure.config.properties.WalletProperties;
 import org.springframework.context.annotation.Configuration;
 
@@ -15,14 +14,12 @@ public class AppConfig {
     private final ApiProperties apiProperties;
     private final IssuerUiProperties issuerUiProperties;
     private final WalletProperties walletProperties;
-    private final TrustServiceProviderForCertificationsProperties trustServiceProviderForCertificationsProperties;
 
-    public AppConfig(ConfigAdapterFactory configAdapterFactory, ApiProperties apiProperties, IssuerUiProperties issuerUiProperties, WalletProperties walletProperties, TrustServiceProviderForCertificationsProperties trustServiceProviderForCertificationsProperties) {
+    public AppConfig(ConfigAdapterFactory configAdapterFactory, ApiProperties apiProperties, IssuerUiProperties issuerUiProperties, WalletProperties walletProperties) {
         this.configAdapter = configAdapterFactory.getAdapter();
         this.apiProperties = apiProperties;
         this.issuerUiProperties = issuerUiProperties;
         this.walletProperties = walletProperties;
-        this.trustServiceProviderForCertificationsProperties = trustServiceProviderForCertificationsProperties;
     }
 
     public String getIssuerApiExternalDomain() {
@@ -47,10 +44,6 @@ public class AppConfig {
 
     public long getCacheLifetimeForVerifiableCredential() {
         return apiProperties.cacheLifetime().verifiableCredential();
-    }
-
-    public String getTrustServiceProviderForCertificationsDid() {
-        return trustServiceProviderForCertificationsProperties.did();
     }
 
 }