+++ src/test/java/es/in2/issuer/infrastructure/config/WebClientConfigTest.java
@@ -2,7 +2,6 @@ package es.in2.issuer.infrastructure.config;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.web.reactive.function.client.WebClient;
 
@@ -10,19 +9,12 @@ import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 class WebClientConfigTest {
 
-    private VerifierConfig verifierConfig;
-
     ApplicationContextRunner context;
 
     @BeforeEach
     void setUp() {
-        // Mocking the VerifierConfig
-        verifierConfig = Mockito.mock(VerifierConfig.class);
-        Mockito.when(verifierConfig.getVerifierExternalDomain()).thenReturn("http://mocked-domain.com");
-
         // Provide the WebClientConfig class and the mock VerifierConfig in the context
         context = new ApplicationContextRunner()
-                .withBean(VerifierConfig.class, () -> verifierConfig)
                 .withUserConfiguration(WebClientConfig.class);
     }
 
@@ -34,11 +26,9 @@ class WebClientConfigTest {
 
             // Assert that both WebClient beans are present
             assertThat(it).hasBean("commonWebClient");
-            assertThat(it).hasBean("oauth2VerifierWebClient");
 
             // Additionally, assert that each bean is of type WebClient
             assertThat(it.getBean("commonWebClient")).isInstanceOf(WebClient.class);
-            assertThat(it.getBean("oauth2VerifierWebClient")).isInstanceOf(WebClient.class);
         });
     }
 }