+++ CHANGELOG.md
@@ -6,8 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v1.2.0
 ### Added
-- Request of obtain M2M tokens from Verifier using LEARmachine VC
-- Validation of M2M tokens
+- Validation of authentication for issuance against the verifier.
 - Verifiable Certifications issuance and sending to response_uri.
 ### Changed
 - List credentials in order from newest to oldest.